requirements.txt
fonttools~=4.40.0
httpx~=0.27.0
Pillow~=9.5.0
playwright~=1.19.0
ruamel.yaml~=0.17.32
tinydb~=4.8.0
websockets~=10.1
